move index at the root of the navigation store
