+++ src/app/store/navigation/index.ts
@@ -1,0 +1,2 @@
+export * as selectedContentEffects from './selected-content/selected-content.effects';
+export { fromSelectedContent } from './selected-content/selected-content.store';
