--- src/app/store/navigation/selected-content/index.ts
@@ -1,2 +1,0 @@
-export * as selectedContentEffects from './selected-content.effects';
-export { fromSelectedContent } from './selected-content.store';
